Add a convenience script to build all pytype targets.
This script will be expanded in a later pass to add options to build
the executables like 'pytype' and 'pytype-single' with C++ logging
enabled.
